% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_assignment_eq.R
\docType{package}
\name{formatEQ-package}
\alias{formatEQ}
\alias{formatEQ-package}
\title{formatEQ: Formats R Code with Preference for Equal Symbol}
\description{
Overrides the formatting function within formatR to change equal
symbols from `<-` to `=`.
}
\details{
The objective of this package is to write a short patch to Yihui Xie's 
formatR package that inverts the behavior of arrow replacement formatting
rule so that instead of replacing \code{=} with \code{<-} the reverse occurs.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

Authors:
\itemize{
  \item Yihui Xie \email{xie@yihui.name}
}

Other contributors:
\itemize{
  \item Kohske Takahashi [contributor]
  \item Ed Lee [contributor]
}

}
