% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_assignment_eq.R
\name{modify_pkg_function}
\alias{modify_pkg_function}
\title{Modify the Definition of a Package Function}
\usage{
modify_pkg_function(fname, fdef, package)
}
\arguments{
\item{fname}{function name}

\item{fdef}{function body definition}

\item{package}{name of package environment}
}
\description{
Places a new definition inside the body of an old function within
a package that is loaded either in the namespace or body.
}
\details{
The implementation of this function is based off of a list serv
conversation response by Henrik Bengtsson.
}
\references{
\url{https://stat.ethz.ch/pipermail/r-help/2008-August/171217.html}
}
